Update Phone Number Tracker
The updated code uses the phone numbers library to parse a user-input phone number and extract the country of origin using the geocoder library. It then uses the langdetect library to detect the language of the input phone number. The program prompts the user to input a phone number in the international format and returns the detected language and country of origin of the phone number. This program can be helpful in verifying the source of phone numbers and detecting potential errors or typos in the input.