Add PDF to DOCX conversion and language prediction functionality
This commit adds functionality to the code to allow for PDF files to be converted to DOCX format using the pdf2docx library. The converted file is then loaded into memory and a language classifier is trained using the converted text. The user is prompted to enter some text, and the trained classifier is used to predict the language of the input text. The predicted language is then printed to the console. The code uses a GUI to allow the user to select the PDF file to convert, the folder to save the converted file, and to enter the input text to predict the language of. The code also logs errors and successful conversions to a log file named 'pdf2docx.log'.