Add GUI for House Price Prediction Model
This commit adds a graphical user interface (GUI) to the house price prediction model created in the previous commit. The GUI was created using the tkinter module and allows users to input values for various features of a house, such as number of bedrooms and bathrooms, and get a predicted price for the house based on the machine learning model. The code for the GUI was added to a new file named "house_price_prediction_GUI.py". The code was tested and works as expected, giving accurate predictions for the input values.